@@ -2,59 +2,80 @@
   "name": "ng2-words",
   "version": "0.0.0",
   "license": "MIT",
+  "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "start": "ng serve --proxy-config proxy.conf.json --port 4202",
+    "start_server": "ng build && node ./express/bin/www",
     "build": "ng build --prod",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "pree2e": "webdriver-manager update --standalone false --gecko false",
+    "e2e": "protractor"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.1",
-    "@angular/common": "^5.1.1",
-    "@angular/compiler": "^5.1.1",
-    "@angular/core": "^5.1.1",
-    "@angular/forms": "^5.1.1",
-    "@angular/http": "^5.1.1",
-    "@angular/platform-browser": "^5.1.1",
-    "@angular/platform-browser-dynamic": "^5.1.1",
-    "@angular/platform-server": "^5.1.1",
-    "@angular/router": "^5.1.1",
-    "bootstrap": "^4.0.0-beta.2",
+    "@angular/animations": "^6.0.9",
+    "@angular/cdk": "^6.4.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/material": "^6.4.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "bootstrap": "^4.1.2",
+    "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
-    "helmet": "^3.9.0",
-    "jquery": "^3.2.1",
-    "lodash": "^4.17.4",
-    "moment": "^2.19.2",
-    "ngx-toastr": "^6.5.0",
-    "popper.js": "^1.13.0",
-    "rxjs": "^5.5.5",
+    "helmet": "^3.12.1",
+    "jquery": "^3.3.1",
+    "lodash": "^4.17.10",
+    "moment": "^2.22.2",
+    "ngx-toastr": "^8.8.0",
+    "passport-google-oauth": "^1.0.0",
+    "popper.js": "^1.14.3",
+    "rxjs": "^6.0.1",
+    "rxjs-compat": "^6.2.2",
     "sjcl": "^1.0.7",
-    "tether": "^1.4.2",
-    "typescript": "2.4.2",
-    "zone.js": "^0.8.14"
+    "tether": "^1.4.4",
+    "ts-helpers": "^1.1.1",
+    "zone.js": "^0.8.26",
+    "async": "latest",
+    "bcrypt-nodejs": "^0.0.3",
+    "bluebird": "^3.5.0",
+    "body-parser": "^1.18.1",
+    "debug": "~2.6.3",
+    "express": "^4.15.4",
+    "express-session": "^1.15.5",
+    "jsonwebtoken": "^8.0.1",
+    "mongodb": "^3.1.1",
+    "mongoose": "^4.11.13",
+    "morgan": "^1.9.0",
+    "nodemailer": "^4.1.1",
+    "owasp-password-strength-test": "^1.3.0",
+    "passport": "latest",
+    "passport-local": "latest",
+    "serve-favicon": "~2.4.2",
+    "validator": "latest"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.3",
-    "@angular/compiler-cli": "^5.1.1",
-    "@angular/language-service": "^5.1.1",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.92",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "webpack": "^3.8.1",
-    "webpack-bundle-analyzer": "^2.9.1"
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.1.5",
+    "@angular/compiler-cli": "^6.0.0",
+    "@types/jasmine": "2.5.38",
+    "@types/node": "^6.0.42",
+    "codelyzer": "^4.4.2",
+    "jasmine-core": "2.5.2",
+    "jasmine-spec-reporter": "2.5.0",
+    "karma": "1.2.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-cli": "^1.0.1",
+    "karma-jasmine": "^1.0.2",
+    "karma-remap-istanbul": "^0.6.0",
+    "protractor": "~4.0.13",
+    "ts-node": "1.2.1",
+    "tslint": "^5.11.0",
+    "typescript": "^2.7.2"
   }
 }
