@@ -2,8 +2,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {LOCALE_ID, NgModule} from '@angular/core';
 import {ReactiveFormsModule} from '@angular/forms';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {HttpClientModule} from '@angular/common/http';
-import {HTTP_INTERCEPTORS} from '@angular/common/http';
+import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
 
 import {AppComponent} from './app.component';
 import {HeaderComponent} from './words-app/components/header/header.component';
@@ -14,19 +13,12 @@ import {UserService} from './words-app/services/user.service';
 import {TimelineService} from './words-app/services/timeline.service';
 import {TextService} from './words-app/services/text.service';
 import {CapitalizePipe} from './words-app/pipes/capitalize.pipe';
-import {AuthInterceptor} from './words-app/services/words-http.service';
+import {AuthInterceptor} from './words-app/services/auth-interceptor.service';
 import {ToastrModule} from 'ngx-toastr';
 import {LoginComponent} from './words-app/layout/login/login.component';
 import {SignupComponent} from './words-app/layout/signup/signup.component';
 
 import {routing} from './app.routes';
-
-import 'rxjs/add/operator/delay';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/throw';
-import 'rxjs/add/operator/map';
 import {Autosize} from './words-app/directives/autosize.directive';
 import {AboutComponent} from './words-app/layout/about/about.component';
 
