@@ -1,49 +1,35 @@
 const express = require('express');
-const path = require('path');
-const favicon = require('serve-favicon');
-const logger = require('morgan');
+const MongoClient = require('mongodb').MongoClient;
 const bodyParser = require('body-parser');
-const expressSession = require('express-session');
-const helmet = require('helmet');
-
-const db = require('./config/mongo');
-const api = require('./routes/api');
-
+const session = require('express-session');
+const passport = require('passport');
+const db = require('./secrets.json').db;
 const app = express();
+const PORT = 5005;
 
-// app.use(favicon(path.join(__dirname, 'dist', 'favicon.ico')));
-app.use(helmet());
-app.use(logger('dev'));
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(expressSession({
-  secret: 'new horse fly sky',
-  saveUninitialized: true,
-  resave: false
+app.use(bodyParser.urlencoded({extended: true}));
+app.use(session({
+    secret: 'new horse fly sky',
+    saveUninitialized: true,
+    resave: true
 }));
+app.use(express.static(require('path').join(__dirname, 'dist')));
 
-app.use(express.static(path.join(__dirname, 'dist')));
+// Passport init
+app.use(passport.initialize({}));
+app.use(passport.session({}));
 
-require('./config/passport')(app);
+MongoClient.connect(
+    `mongodb://${db.dbuser}:${db.dbpassword}@${db.dbhost}/${db.dbname}`,
+    {useNewUrlParser: true},
+    (err, database) => {
+        if (err) return console.log(err);
+        require('./routes')(app, database);
+        app.listen(PORT, () => {
+            console.log('We are live on http://localhost:' + PORT);
+        });
+    });
 
-app.use('/api', api);
-
-// catch 404 and forward to error handler
-app.use(function (req, res, next) {
-  const err = new Error('Not Found ' + req.path);
-  err.status = 404;
-  next(err);
-});
-
-// error handler
-app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  console.log(res.locals);
-});
 
 module.exports = app;
